@@ -1,5 +1,6 @@
 package com.timeflow.ui.screens
 
+import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
@@ -15,36 +16,52 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
+import com.timeflow.R
 import com.timeflow.model.EventType
 import com.timeflow.model.TimelineEvent
-import com.timeflow.ui.components.EventDialog
-import com.timeflow.ui.components.QuickAddEventForm
+import com.timeflow.ui.components.EventTypeChip
+import com.timeflow.ui.screens.EventDetailDialog
 import com.timeflow.viewmodel.TimelineViewModel
-import java.text.SimpleDateFormat
-import java.util.*
+import java.time.LocalDateTime
+import java.time.ZoneId
+import java.time.format.DateTimeFormatter
+import java.util.Date
 
-/**
+
+    /**
  * 时间轴屏幕，显示事件列表、搜索和过滤功能
- */
+     */
 @OptIn(ExperimentalMaterial3Api::class)
+
 @Composable
+
+
 fun TimelineScreen(viewModel: TimelineViewModel) {
-    val events = viewModel.getFilteredEvents()
-    val searchTerm = viewModel.searchTerm
+
+    LaunchedEffect(Unit) {
+        viewModel.loadEvents()
+    }
+
+    val events by viewModel.events.collectAsState(initial = emptyList())    
+    val searchTerm by viewModel.searchTerm.collectAsState()
     val selectedEventType = viewModel.selectedEventType
     val editingEvent = viewModel.editingEvent
-    
+
+    var showEventDetailDialog by remember { mutableStateOf(false) }
+    var selectedEvent by remember { mutableStateOf<TimelineEvent?>(null) }
+    var showSettingDialog by remember { mutableStateOf(false) }
+    var isDarkTheme by remember { mutableStateOf(false) }
+
     var showAddEventDialog by remember { mutableStateOf(false) }
     var showEditEventDialog by remember { mutableStateOf(false) }
     var showDeleteConfirmDialog by remember { mutableStateOf(false) }
     var eventToDelete by remember { mutableStateOf<TimelineEvent?>(null) }
-    var searchExpanded by remember { mutableStateOf(false) }
-    var showFilterMenu by remember { mutableStateOf(false) }
+    var searchExpanded by remember { mutableStateOf(false) } 
     
     // 当编辑事件变化时更新对话框状态
     LaunchedEffect(editingEvent) {
@@ -55,13 +72,25 @@ fun TimelineScreen(viewModel: TimelineViewModel) {
         topBar = {
             TopAppBar(
                 title = { Text("TimeFlow") },
-                actions = {
+                actions = { // 搜索按钮
+                    IconButton(onClick = { showSettingDialog = !showSettingDialog }) {
+                        Icon(Icons.Outlined.Settings, contentDescription = "设置")
+                    }
+                    // 设置对话框
+                    if (showSettingDialog) {
+                        SettingDialog(onDismiss = { showSettingDialog = false }, onThemeChange = {isDarkTheme = it}, isDarkTheme = isDarkTheme)
+                    }
+
                     // 搜索按钮
-                    IconButton(onClick = { searchExpanded = true }) {
+                    IconButton(onClick = {
+                        searchExpanded = !searchExpanded
+                    }) {
                         Icon(Icons.Default.Search, contentDescription = "搜索")
                     }
                     // 过滤按钮
-                    IconButton(onClick = { showFilterMenu = true }) {
+                    var showFilterMenu by remember { mutableStateOf(false) }
+                    IconButton(onClick = {
+                        showFilterMenu = !showFilterMenu }) {
                         Icon(Icons.Default.FilterList, contentDescription = "过滤")
                     }
                 }
@@ -99,12 +128,12 @@ fun TimelineScreen(viewModel: TimelineViewModel) {
                     }
                 ) {}
             }
-            
+
             // 事件类型过滤器
             if (showFilterMenu) {
                 FilterDropdownMenu(
                     selectedEventType = selectedEventType,
-                    onEventTypeSelected = { 
+                    onEventTypeSelected = {
                         viewModel.setEventTypeFilter(it)
                         showFilterMenu = false
                     },
@@ -112,19 +141,36 @@ fun TimelineScreen(viewModel: TimelineViewModel) {
                 )
             }
             
-            // 时间轴列表
-            LazyColumn(
-                modifier = Modifier
+            
+            // 时间轴列表 
+            Box(modifier = Modifier.fillMaxSize()) {
+                Canvas(modifier = Modifier
                     .fillMaxSize()
-                    .padding(horizontal = 16.dp)
-            ) {
-                items(events, key = { it.id }) { event ->
-                    TimelineItem(
-                        event = event,
-                        onEditClick = { viewModel.setEditingEvent(event) },
-                        onDeleteClick = {
-                            eventToDelete = event
-                            showDeleteConfirmDialog = true
+                    .width(1.dp)
+                    .align(Alignment.CenterStart)
+                    .padding(start = 40.dp)) {
+                    drawLine(color = MaterialTheme.colorScheme.outline, start = androidx.compose.ui.geometry.Offset(0f, 0f), end = androidx.compose.ui.geometry.Offset(0f, size.height)) 
+                } 
+                LazyColumn(modifier = Modifier
+                        .fillMaxSize()
+                        .padding(horizontal = 16.dp)) {
+                    val groupedEvents = events.groupBy { LocalDate.ofInstant(it.timestamp.toInstant(), ZoneId.systemDefault()) }
+                    groupedEvents.forEach { (date, eventsInDay) ->
+                        item(key = date.toString()) { DateHeader(date = date)}
+                    }
+                    groupedEvents.forEach { (_, eventsInDay) ->
+                        items(eventsInDay, key = { it.id }) { event ->
+                            TimelineItem( 
+                            event = event,
+                                onEventClick = {  selectedEvent = event
+                                    showEventDetailDialog = true },
+                                onEditClick = { viewModel.setEditingEvent(event) },
+
+                                onDeleteClick = { eventToDelete = event
+                                    showDeleteConfirmDialog = true }
+
+                            )
+                         }
                         }
                     )
                 }
@@ -139,11 +185,10 @@ fun TimelineScreen(viewModel: TimelineViewModel) {
                     horizontalAlignment = Alignment.CenterHorizontally,
                     verticalArrangement = Arrangement.Center
                 ) {
-                    Icon(
-                        imageVector = Icons.Outlined.EventNote,
-                        contentDescription = null,
-                        modifier = Modifier.size(64.dp),
-                        tint = MaterialTheme.colorScheme.primary.copy(alpha = 0.5f)
+                    androidx.compose.foundation.Image(
+                        painter = painterResource(id = R.drawable.empty),
+                        contentDescription = "Empty Timeline",
+                        modifier = Modifier.size(200.dp)
                     )
                     Spacer(modifier = Modifier.height(16.dp))
                     Text(
@@ -158,6 +203,7 @@ fun TimelineScreen(viewModel: TimelineViewModel) {
                 }
             }
         }
+        
     }
     
     // 添加事件对话框
@@ -165,8 +211,8 @@ fun TimelineScreen(viewModel: TimelineViewModel) {
         EventDialog(
             event = null,
             onDismiss = { showAddEventDialog = false },
-            onSave = { eventType, description, imageUrl, attachment ->
-                viewModel.addEvent(eventType, description, imageUrl, attachment)
+            onSave = { eventType, description, imageUrl, attachment, eventTimestamp ->
+                viewModel.addEvent(eventType, description, imageUrl, attachment, eventTimestamp)
                 showAddEventDialog = false
             }
         )
@@ -178,15 +224,16 @@ fun TimelineScreen(viewModel: TimelineViewModel) {
             event = editingEvent,
             onDismiss = { 
                 viewModel.setEditingEvent(null)
-                showEditEventDialog = false 
+                        showEditEventDialog = false
             },
-            onSave = { eventType, description, imageUrl, attachment ->
+            onSave = { eventType, description, imageUrl, attachment, eventTimestamp ->
                 val updatedEvent = editingEvent.copy(
                     eventType = eventType,
                     title = TimelineEvent.deriveTitle(description),
                     description = description,
                     imageUrl = imageUrl,
-                    attachment = attachment
+                    attachment = attachment,
+                    timestamp = eventTimestamp
                 )
                 viewModel.updateEvent(updatedEvent)
                 showEditEventDialog = false
@@ -226,25 +273,53 @@ fun TimelineScreen(viewModel: TimelineViewModel) {
             }
         )
     }
+    
+    // 事件详情对话框
+    if (showEventDetailDialog && selectedEvent != null) {
+        EventDetailDialog(
+            event = selectedEvent!!,
+            onDismiss = {
+                showEventDetailDialog = false
+                selectedEvent = null
+
+            }
+        })
+    }
 }
 
 /**
  * 时间轴项目组件
  */
-@Composable
-fun TimelineItem(
-    event: TimelineEvent,
-    onEditClick: () -> Unit,
-    onDeleteClick: () -> Unit
-) {
-    val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm", Locale.getDefault())
-    
-    Card(
+@Composable 
+fun TimelineItem(event: TimelineEvent, onEventClick: () -> Unit, onEditClick: () -> Unit, onDeleteClick: () -> Unit) { 
+
+    val dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")
+    val localDateTime: LocalDateTime =
+        event.timestamp.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime()
+    Row(modifier = Modifier.fillMaxWidth()) {
+        Canvas(
         modifier = Modifier
+                .clickable {
+                    onEventClick() 
+                }
+                .padding(top = 16.dp), onDraw = {
+            drawCircle(
+                color = MaterialTheme.colorScheme.primary,
+                radius = 8.dp.toPx(),
+                center = androidx.compose.ui.geometry.Offset(
+                    size.width / 2 - 40.dp.toPx(), 0f
+                )
+            )
+            drawLine(color = MaterialTheme.colorScheme.outline, start = androidx.compose.ui.geometry.Offset(size.width / 2 - 40.dp.toPx(), 0f), end = androidx.compose.ui.geometry.Offset(size.width / 2 - 40.dp.toPx(), size.height)) 
+        })
+
+        Card(
+
+            modifier = Modifier
             .fillMaxWidth()
-            .padding(vertical = 8.dp),
-        shape = RoundedCornerShape(12.dp),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
+            .padding(vertical = 8.dp).clickable { onEventClick() },
+                shape = RoundedCornerShape(12.dp),
+                elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
     ) {
         Column(modifier = Modifier.padding(16.dp)) {
             // 头部：类型图标、标题和时间
@@ -280,7 +355,7 @@ fun TimelineItem(
                 }
                 // 时间
                 Text(
-                    text = dateFormat.format(event.timestamp),
+                    text = dateTimeFormatter.format(localDateTime),
                     style = MaterialTheme.typography.bodySmall,
                     color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.6f)
                 )
@@ -333,7 +408,7 @@ fun TimelineItem(
                     )
                 }
             }
-            
+
             Spacer(modifier = Modifier.height(12.dp))
             
             // 操作按钮
@@ -349,6 +424,7 @@ fun TimelineItem(
                         tint = MaterialTheme.colorScheme.primary
                     )
                 }
+                Spacer(modifier = Modifier.width(8.dp))
                 // 删除按钮
                 IconButton(onClick = onDeleteClick) {
                     Icon(
@@ -359,6 +435,22 @@ fun TimelineItem(
                 }
             }
         }
+    }
+}
+
+
+@Composable
+fun DateHeader(date: LocalDate) {
+    val dateFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd EEEE")
+    val formattedDate = date.format(dateFormatter)
+    Row(modifier = Modifier.fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Spacer(modifier = Modifier.width(80.dp))
+        Text( 
+            text = formattedDate,
+            style = MaterialTheme.typography.headlineSmall
+        )
     }
 }
 
@@ -419,13 +511,14 @@ fun FilterDropdownMenu(
 }
 
 /**
- * 获取事件类型对应的图标
+ * Get the correct icon from the eventType
  */
 @Composable
 fun getEventTypeIcon(eventType: EventType) = when (eventType) {
-    EventType.NOTE -> Icons.Default.Note
-    EventType.TODO -> Icons.Default.CheckBox
-    EventType.SCHEDULE -> Icons.Default.Event
+    EventType.NOTE -> Icons.Outlined.Note
+    TODO -> Icons.Outlined.CheckBox
+    SCHEDULE -> Icons.Outlined.Event
+    MEMO -> Icons.Outlined.EditNote
 }
 
 /**
@@ -436,4 +529,5 @@ fun getEventTypeColor(eventType: EventType) = when (eventType) {
     EventType.NOTE -> MaterialTheme.colorScheme.tertiary
     EventType.TODO -> MaterialTheme.colorScheme.primary
     EventType.SCHEDULE -> MaterialTheme.colorScheme.secondary
+    EventType.MEMO -> MaterialTheme.colorScheme.surfaceTint
 }
