@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.timeflow.ui.theme.TimeFlowTheme
@@ -17,13 +18,18 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            TimeFlowTheme {
+            val viewModel: TimelineViewModel = viewModel()
+            val context = LocalContext.current
+            val isDarkTheme = viewModel.isDarkTheme
+            TimeFlowTheme(darkTheme = isDarkTheme) {
                 Surface(
-                    modifier = Modifier.fillMaxSize(),
+                    modifier = Modifier.fillMaxSize() ,
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    val viewModel: TimelineViewModel = viewModel()
-                    TimelineScreen(viewModel = viewModel)
+                    TimelineScreen(
+                        viewModel = viewModel,
+                        isDarkTheme = isDarkTheme
+                    )
                 }
             }
         }
@@ -34,7 +40,7 @@ class MainActivity : ComponentActivity() {
 @Composable
 fun DefaultPreview() {
     TimeFlowTheme {
-        val viewModel = TimelineViewModel()
-        TimelineScreen(viewModel = viewModel)
+        val viewModel: TimelineViewModel = viewModel()
+        TimelineScreen(viewModel = viewModel, isDarkTheme = false)
     }
 }
