@@ -9,19 +9,25 @@ class Converters {
     fun fromTimestamp(value: Long?): Date? {
         return value?.let { Date(it) }
     }
-    
+
     @TypeConverter
     fun dateToTimestamp(date: Date?): Long? {
         return date?.time
     }
-    
+
     @TypeConverter
     fun fromEventType(value: EventType): String {
         return value.name
     }
-    
+
     @TypeConverter
     fun toEventType(value: String): EventType {
-        return EventType.valueOf(value)
+        return when (value) {
+            "NOTE" -> EventType.NOTE
+            "TODO" -> EventType.TODO
+            "SCHEDULE" -> EventType.SCHEDULE
+            "MEMO" -> EventType.MEMO
+            else -> EventType.NOTE
+        }
     }
-} 
+}
